job_portal/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.0.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
